src/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
faker
pandas
dotenv
pyodbc
azure-storage-blob